server/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.9.0
python-dotenv>=1.0.0
openai>=1.0.0
tiktoken>=0.5.0

# LangChain framework dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
